requirements.txt
Pillow==8.4.0
psutil==5.9.5
PyMySQL==1.0.3
pyserial==3.5
APScheduler==3.10.1